requirements.txt
requests
PIL
bs4
fpdf2